backend/requirements.txt
flask==3.0.2
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.38
alembic==1.14.1
PyMySQL==1.1.1
pydantic==2.10.6
email-validator==2.2.0
argon2-cffi==23.1.0
pytest==8.3.4
PyJWT==2.10.1
